Implemented company owner endpoint (#103)

+++ src/app/modules/corporate-contributor/component/configure-cla-manager-modal/configure-cla-manager-modal.component.ts
@@ -1,7 +1,7 @@
 // Copyright The Linux Foundation and each contributor to CommunityBridge.
 // SPDX-License-Identifier: MIT
 
-import { Component, TemplateRef, ViewChild, EventEmitter, Output } from '@angular/core';
+import { Component, TemplateRef, ViewChild, EventEmitter, Output, OnInit } from '@angular/core';
 import { ClaContributorService } from 'src/app/core/services/cla-contributor.service';
 import { AuthService } from 'src/app/shared/services/auth.service';
 import { AppSettings } from 'src/app/config/app-settings';
@@ -15,7 +15,7 @@ import { AlertService } from 'src/app/shared/services/alert.service';
   templateUrl: './configure-cla-manager-modal.component.html',
   styleUrls: ['./configure-cla-manager-modal.component.scss']
 })
-export class ConfigureClaManagerModalComponent {
+export class ConfigureClaManagerModalComponent implements OnInit {
   @ViewChild('errorModal') errorModal: TemplateRef<any>;
   @ViewChild('warningModal') warningModal: TemplateRef<any>;
   @Output() showCloseBtnEmitter: EventEmitter<any> = new EventEmitter<any>();
@@ -24,6 +24,7 @@ export class ConfigureClaManagerModalComponent {
   message: string;
   company: OrganizationModel;
   hasCLAManagerDesignee: boolean;
+  hasCompanyOwner: boolean;
 
   constructor(
     private claContributorService: ClaContributorService,
@@ -33,17 +34,21 @@ export class ConfigureClaManagerModalComponent {
     private alertService: AlertService
   ) {
     this.hasCLAManagerDesignee = false;
+    this.hasCompanyOwner = false;
     this.showCloseBtnEmitter.emit(false);
-
     setTimeout(() => {
       this.manageAuthRedirection();
     }, 100);
   }
 
+  ngOnInit() {
+    this.company = JSON.parse(this.storageService.getItem(AppSettings.SELECTED_COMPANY));
+  }
+
   manageAuthRedirection() {
     const actionType = JSON.parse(this.storageService.getItem(AppSettings.ACTION_TYPE));
     if (actionType === AppSettings.SIGN_CLA) {
-      this.addAsCLAManagerDesignee();
+      this.addContributorAsDesigneeAndOwner();
     } else {
       this.validateUserLFID();
     }
@@ -51,8 +56,8 @@ export class ConfigureClaManagerModalComponent {
 
   validateUserLFID() {
     if (this.claContributorService.getUserLFID()) {
-      if (this.claContributorService.hasTokenValid()) {
-        this.addAsCLAManagerDesignee();
+      if (this.claContributorService.hasTokenValid() && this.authService.isAuthenticated()) {
+        this.addContributorAsDesigneeAndOwner();
       } else {
         this.redirectToAuth0();
       }
@@ -64,20 +69,25 @@ export class ConfigureClaManagerModalComponent {
     }
   }
 
-  addAsCLAManagerDesignee() {
-    this.hasCLAManagerDesignee = false;
-    this.company = JSON.parse(this.storageService.getItem(AppSettings.SELECTED_COMPANY));
-    const projectId = JSON.parse(this.storageService.getItem(AppSettings.PROJECT_ID));
+  addContributorAsDesigneeAndOwner() {
     const data = {
       userEmail: this.claContributorService.getUserPublicEmail()
     };
+    this.addAsCLAManagerDesignee(data);
+    this.addAsCompanyOwner(data);
+  }
+
+  addAsCLAManagerDesignee(data: any) {
+    const projectId = JSON.parse(this.storageService.getItem(AppSettings.PROJECT_ID));
     this.claContributorService.addAsCLAManagerDesignee(this.company.companyExternalID, projectId, data).subscribe(
       () => {
+        this.hasCLAManagerDesignee = true;
         this.proccedToCorporateConsole();
       },
       (exception) => {
         if (exception.status === 409) {
           // User has already CLA manager designee.
+          this.hasCLAManagerDesignee = true;
           this.proccedToCorporateConsole();
         } else {
           this.title = 'Request Failed';
@@ -89,10 +99,26 @@ export class ConfigureClaManagerModalComponent {
     );
   }
 
+  addAsCompanyOwner(data: any) {
+    this.claContributorService.addAsCompanyOwner(this.company.companyExternalID, data).subscribe(
+      () => {
+        this.hasCompanyOwner = true;
+        this.proccedToCorporateConsole();
+      },
+      (exception) => {
+        this.title = 'Request Failed';
+        this.storageService.removeItem(AppSettings.ACTION_TYPE);
+        this.message = exception.error.Message;
+        this.openDialog(this.errorModal);
+      }
+    );
+  }
+
   proccedToCorporateConsole() {
-    this.storageService.removeItem(AppSettings.ACTION_TYPE);
-    this.hasCLAManagerDesignee = true;
-    this.showCloseBtnEmitter.emit(true);
+    if (this.hasCLAManagerDesignee && this.hasCompanyOwner) {
+      this.storageService.removeItem(AppSettings.ACTION_TYPE);
+      this.showCloseBtnEmitter.emit(true);
+    }
   }
 
   onClickProceedBtn() {
@@ -103,24 +129,24 @@ export class ConfigureClaManagerModalComponent {
   }
 
   onClickProccedModalBtn() {
-    if (!this.hasCLAManagerDesignee) {
-      this.redirectToAuth0();
-    } else {
-      this.modalService.dismissAll();
-      const hasGerrit = JSON.parse(this.storageService.getItem(AppSettings.HAS_GERRIT));
-      const flashMsg = 'Your ' + (hasGerrit ? 'Gerrit' : 'GitHub') + ' session has been preserved in the current tab so that you can always come back to it after completing CLA signing';
-      this.alertService.success(flashMsg);
-      setTimeout(() => {
-        this.storageService.removeItem(AppSettings.ACTION_TYPE);
-        const corporateUrl = this.claContributorService.getLFXCorporateURL();
-        window.open(corporateUrl, '_blank');
-      }, 4500);
+    this.modalService.dismissAll();
+    const hasGerrit = JSON.parse(this.storageService.getItem(AppSettings.HAS_GERRIT));
+    const flashMsg = 'Your ' + (hasGerrit ? 'Gerrit' : 'GitHub') + ' session has been preserved in the current tab so that you can always come back to it after completing CLA signing';
+    this.alertService.success(flashMsg);
+    setTimeout(() => {
+      this.storageService.removeItem(AppSettings.ACTION_TYPE);
+      const corporateUrl = this.claContributorService.getLFXCorporateURL();
+      window.open(corporateUrl, '_blank');
+    }, 4500);
 
-      setTimeout(() => {
-        const redirectUrl = JSON.parse(this.storageService.getItem(AppSettings.REDIRECT));
+    setTimeout(() => {
+      const redirectUrl = JSON.parse(this.storageService.getItem(AppSettings.REDIRECT));
+      if (redirectUrl) {
         window.open(redirectUrl, '_self');
-      }, 5000);
-    }
+      } else {
+        this.alertService.error('Error occured while redirection please confirm you come to the contributor console by following proper steps.');
+      }
+    }, 5000);
   }
 
   redirectToAuth0() {
